requirements.txt
torch
torchvision
wandb
streamlit
matplotlib
seaborn
pandas
numpy
python-dotenv
tqdm
scikit-learn
plotly
